A table of official country names, ISO-3 codes and other metadata

Provides official codes taken from the United Nations official sources. This data is used by the functions provided within this package to validate user parameters and add both product and country text columns to the data, therefore reducing the number of API calls and the time to generate the requested data.

ots_attributes_countries

Format

A data frame with 249 observations on the following 6 variables.

country_iso

ISO code of the country (e.g. "chl" means Chile)

country_name_english

Country name (e.g. Germany)

country_fullname_english

Country name with indications (e.g. Germany (former Federal Republic of Germany until 1990))

continent_id

Numeric id of the continent where the country belongs to

continent

Continent where the country belongs to

eu28_member

Dummy variable such that 1 means "belongs to EU-28 group" and 0 otherwise

\

Examples

ots_attributes_countries

#> # A tibble: 249 x 6
#>    country_iso country_name_en… country_fullnam… continent_id continent
#>    <chr>       <chr>            <chr>                   <int> <chr>    
#>  1 afg         Afghanistan      Afghanistan                 1 Asia     
#>  2 alb         Albania          Albania                     2 Europe   
#>  3 dza         Algeria          Algeria                     3 Africa   
#>  4 asm         American Samoa   American Samoa              4 Oceania  
#>  5 and         Andorra          Andorra                     2 Europe   
#>  6 ago         Angola           Angola                      3 Africa   
#>  7 aia         Anguilla         Anguilla                    5 Americas 
#>  8 atg         Antigua and Bar… Antigua and Bar…            5 Americas 
#>  9 arg         Argentina        Argentina                   5 Americas 
#> 10 arm         Armenia          Armenia                     1 Asia     
#> # … with 239 more rows, and 1 more variable: eu28_member <int>