A table of communities codes, names and colours associated to products

Provides communities from The Atlas of Economic Complexity (Harvard), and those can be useful for grouping into broader categories than those the official HS groups.

ots_communities

Format

A data frame with 6275 observations on the following 4 variables.

product_code

Code of every product (e.g. 0101)

community_code

Community code (e.g. 01)

community_name

Community name (e.g. Animal Products)

community_color

Colour hex code (e.g. #74c0e2)

Examples

ots_communities

#> # A tibble: 6,275 x 4
#>    product_code community_code community_name  community_color
#>    <chr>        <chr>          <chr>           <chr>          
#>  1 0101         01             Animal Products #74c0e2        
#>  2 010110       01             Animal Products #74c0e2        
#>  3 010190       01             Animal Products #74c0e2        
#>  4 0102         01             Animal Products #74c0e2        
#>  5 010210       01             Animal Products #74c0e2        
#>  6 010290       01             Animal Products #74c0e2        
#>  7 0103         01             Animal Products #74c0e2        
#>  8 010310       01             Animal Products #74c0e2        
#>  9 010391       01             Animal Products #74c0e2        
#> 10 010392       01             Animal Products #74c0e2        
#> # … with 6,265 more rows