String matching of official product names and Harmonized System (HS) codes according to the United Nations nomenclature

Takes a text string and searches within the package data for all matching product codes in the context of valid API product codes.

ots_product_code(productname = NULL, productgroup = NULL)

Arguments

productname
A text string such as "Animals", "COPPER" or "fruits".

productgroup
A text string such as "meat", "FISH" or "Dairy".

Value

A tibble with all possible matches (no uppercase distinction) showing the product name and product code

Examples

ots_product_code(productname = "ANIMALS ")

#>     product_code
#>  1:         0102
#>  2:         0106
#>  3:         0201
#>  4:         0202
#>  5:         0206
#>  6:         0504
#>  7:         0511
#>  8:         1502
#>  9:         4101
#> 10:         4104
#> 11:         4106
#> 12:         4107
#> 13:         4113
#> 14:           01
#>                                                                                                                                                                                                    product_fullname_english
#>  1:                                                                                                                                                                                                    Bovine animals; live
#>  2:                                                                                                                                                                                     Animals, n.e.c. in chapter 01; live
#>  3:                                                                                                                                                                                Meat of bovine animals; fresh or chilled
#>  4:                                                                                                                                                                                          Meat of bovine animals; frozen
#>  5:                                                                                                          Edible offal of bovine animals, swine, sheep, goats, horses, asses, mules or hinnies; fresh, chilled or frozen
#>  6:                                                                           Guts, bladders and stomachs of animals (other than fish); whole and pieces thereof, fresh, chilled, frozen, salted, in brine, dried or smoked
#>  7:                                                                                                        Animal products not elsewhere specified or included; dead animals of chapter 1 or 3, unfit for human consumption
#>  8:                                                                                                                                            Fats of bovine animals, sheep or goats; other than those of heading no. 1503
#>  9: Raw hides and skins of bovine (including buffalo) or equine animals (fresh, salted, dried, limed, pickled, otherwise preserved but not tanned, parchment dressed or further prepared), whether or not dehaired or split
#> 10:                                                                        Tanned or crust hides and skins of bovine (including buffalo) or equine animals, without hair on, whether or not split, but not further prepared
#> 11:                                                                                               Tanned or crust hides and skins of other animals, without wool or hair on, whether or not split, but not further prepared
#> 12:    Leather further prepared after tanning or crusting, including parchment-dressed leather, of bovine (including buffalo) or equine animals, without hair on, whether or not split, other than leather of heading 41.14
#> 13:              Leather further prepared after tanning or crusting, including parchment-dressed leather, of animals (other than ovine), without wool or hair on, whether or not split, other than leather of heading 41.14
#> 14:                                                                                                                                                                          Alias for all codes in the group Animals; live
#>     type_product
#>  1:      ANIMALS
#>  2:      ANIMALS
#>  3:      ANIMALS
#>  4:      ANIMALS
#>  5:      ANIMALS
#>  6:      ANIMALS
#>  7:      ANIMALS
#>  8:      ANIMALS
#>  9:      ANIMALS
#> 10:      ANIMALS
#> 11:      ANIMALS
#> 12:      ANIMALS
#> 13:      ANIMALS
#> 14:      ANIMALS

ots_product_code(productgroup = "  fish")

#>    group_code
#> 1:         03
#> 2:         16
#>                                                                      group_fullname_english
#> 1:                           Fish and crustaceans, molluscs and other aquatic invertebrates
#> 2: Meat, fish or crustaceans, molluscs or other aquatic invertebrates; preparations thereof
#>    type_group
#> 1:       fish
#> 2:       fish

ots_product_code(productname = "Milk", productgroup = "Dairy")

#>    group_code
#> 1:         04
#> 2:         04
#> 3:         04
#> 4:         04
#> 5:         04
#>                                                                                              group_fullname_english
#> 1: Dairy produce; birds' eggs; natural honey; edible products of animal origin, not elsewhere specified or included
#> 2: Dairy produce; birds' eggs; natural honey; edible products of animal origin, not elsewhere specified or included
#> 3: Dairy produce; birds' eggs; natural honey; edible products of animal origin, not elsewhere specified or included
#> 4: Dairy produce; birds' eggs; natural honey; edible products of animal origin, not elsewhere specified or included
#> 5: Dairy produce; birds' eggs; natural honey; edible products of animal origin, not elsewhere specified or included
#>    product_code
#> 1:         0401
#> 2:         0402
#> 3:         0403
#> 4:         0404
#> 5:         0405
#>                                                                                                                                                                                product_fullname_english
#> 1:                                                                                                               Milk and cream; not concentrated nor containing added sugar or other sweetening matter
#> 2:                                                                                                                    Milk and cream; concentrated or containing added sugar or other sweetening matter
#> 3: Buttermilk, curdled milk and cream, yoghurt, kephir, fermented or acidified milk or cream, whether or not concentrated, containing added sugar, sweetening matter, flavoured or added fruit or cocoa
#> 4:                                     Whey and products consisting of natural milk constituents; whether or not containing added sugar or other sweetening matter, not elsewhere specified or included
#> 5:                                                                                                                                      Butter and other fats and oils derived from milk; dairy spreads
#>    type_name type_group
#> 1:      Milk      Dairy
#> 2:      Milk      Dairy
#> 3:      Milk      Dairy
#> 4:      Milk      Dairy
#> 5:      Milk      Dairy