String matching of official product section names and product section codes

Takes a text string and searches within the package data for all matching product communities in the context of valid API product communities

ots_product_section(productsection = NULL)

Arguments

productsection
A text string such as "animals", or "FOODSTUFFS".

Value

A tibble with all possible matches (no uppercase distinction) showing the section name and section code

Examples

ots_product_section(productsection = "  Animals")

#>     section_code      section_fullname_english product_code type_section
#>  1:           01 Live Animals; Animal Products         0101      Animals
#>  2:           01 Live Animals; Animal Products         0102      Animals
#>  3:           01 Live Animals; Animal Products         0103      Animals
#>  4:           01 Live Animals; Animal Products         0104      Animals
#>  5:           01 Live Animals; Animal Products         0105      Animals
#>  6:           01 Live Animals; Animal Products         0106      Animals
#>  7:           01 Live Animals; Animal Products         0201      Animals
#>  8:           01 Live Animals; Animal Products         0202      Animals
#>  9:           01 Live Animals; Animal Products         0203      Animals
#> 10:           01 Live Animals; Animal Products         0204      Animals
#> 11:           01 Live Animals; Animal Products         0205      Animals
#> 12:           01 Live Animals; Animal Products         0206      Animals
#> 13:           01 Live Animals; Animal Products         0207      Animals
#> 14:           01 Live Animals; Animal Products         0208      Animals
#> 15:           01 Live Animals; Animal Products         0209      Animals
#> 16:           01 Live Animals; Animal Products         0210      Animals
#> 17:           01 Live Animals; Animal Products         0301      Animals
#> 18:           01 Live Animals; Animal Products         0302      Animals
#> 19:           01 Live Animals; Animal Products         0303      Animals
#> 20:           01 Live Animals; Animal Products         0304      Animals
#> 21:           01 Live Animals; Animal Products         0305      Animals
#> 22:           01 Live Animals; Animal Products         0306      Animals
#> 23:           01 Live Animals; Animal Products         0307      Animals
#> 24:           01 Live Animals; Animal Products         0401      Animals
#> 25:           01 Live Animals; Animal Products         0402      Animals
#> 26:           01 Live Animals; Animal Products         0403      Animals
#> 27:           01 Live Animals; Animal Products         0404      Animals
#> 28:           01 Live Animals; Animal Products         0405      Animals
#> 29:           01 Live Animals; Animal Products         0406      Animals
#> 30:           01 Live Animals; Animal Products         0407      Animals
#> 31:           01 Live Animals; Animal Products         0408      Animals
#> 32:           01 Live Animals; Animal Products         0409      Animals
#> 33:           01 Live Animals; Animal Products         0410      Animals
#> 34:           01 Live Animals; Animal Products         0501      Animals
#> 35:           01 Live Animals; Animal Products         0502      Animals
#> 36:           01 Live Animals; Animal Products         0504      Animals
#> 37:           01 Live Animals; Animal Products         0505      Animals
#> 38:           01 Live Animals; Animal Products         0506      Animals
#> 39:           01 Live Animals; Animal Products         0507      Animals
#> 40:           01 Live Animals; Animal Products         0508      Animals
#> 41:           01 Live Animals; Animal Products         0510      Animals
#> 42:           01 Live Animals; Animal Products         0511      Animals
#>     section_code      section_fullname_english product_code type_section

ots_product_section(productsection = "FABRIC ")

#> Empty data.table (0 rows and 4 cols): section_code,section_fullname_english,product_code,type_section