A table of official product names, Harmonized System (HS) codes and corresponding groups

Provides official codes taken from the United Nations official sources. This data is provided as a reference as it's not used by the functions provided within this package, but it can be useful to subset the data obtained from the API.

ots_products

Format

A data frame with 6373 observations on the following 4 variables.

product_code

Code of every product (e.g. 0101)

product_fullname_english

HS product names (e.g. 'Horses, asses, mules and hinnies; live')

group_code

HS group code (e.g. 01)

group_name

HS group name (e.g. 'Animals; live')

Examples

ots_products

#> # A tibble: 6,373 x 4
#>    product_code product_fullname_english                  group_code group_name 
#>    <chr>        <chr>                                     <chr>      <chr>      
#>  1 0101         Horses, asses, mules and hinnies; live    01         Animals; l…
#>  2 010110       Horses, asses, mules and hinnies; live, … 01         Animals; l…
#>  3 010190       Horses, asses, mules and hinnies; live, … 01         Animals; l…
#>  4 0102         Bovine animals; live                      01         Animals; l…
#>  5 010210       Bovine animals; live, pure-bred breeding… 01         Animals; l…
#>  6 010290       Bovine animals; live, other than pure-br… 01         Animals; l…
#>  7 0103         Swine; live                               01         Animals; l…
#>  8 010310       Swine; live, pure-bred breeding animals   01         Animals; l…
#>  9 010391       Swine; live, (other than pure-bred breed… 01         Animals; l…
#> 10 010392       Swine; live, (other than pure-bred breed… 01         Animals; l…
#> # … with 6,363 more rows