A table of sections codes and official names associated to products

Provides section names from the Harmonized System.

ots_sections_names

Format

A data frame with 2 observations on the following 2 variables.

section_code

Section code (e.g. 01)

section_fullname_english

Section name (e.g. Live Animals; Animal Products)

Examples

ots_sections_names

#>     section_code
#>  1:           01
#>  2:           02
#>  3:           03
#>  4:           04
#>  5:           05
#>  6:           06
#>  7:           07
#>  8:           08
#>  9:           09
#> 10:           10
#> 11:           11
#> 12:           12
#> 13:           13
#> 14:           14
#> 15:           15
#> 16:           16
#> 17:           17
#> 18:           18
#> 19:           19
#> 20:           20
#> 21:           21
#> 22:           22
#>     section_code
#>                                                                                                                                                                                        section_fullname_english
#>  1:                                                                                                                                                                               Live Animals; Animal Products
#>  2:                                                                                                                                                                                          Vegetable Products
#>  3:                                                                                              Animal Or Vegetable Fats And Oils And Their Cleavage Products; Prepared Edible Fats; Animal Or Vegetable Waxes
#>  4:                                                                                                           Prepared Foodstuffs; Beverages, Spirits And Vinegar; Tobacco And Manufactured Tobacco Substitutes
#>  5:                                                                                                                                                                                            Mineral Products
#>  6:                                                                                                                                                               Products Of The Chemical Or Allied Industries
#>  7:                                                                                                                                                  Plastics And Articles Thereof; Rubber And Articles Thereof
#>  8:                         Raw Hides And Skins, Leather, Furskins And Articles Thereof; Saddlery And Harness; Travel Goods, Handbags And Similar Containers; Articles Of Animal Gut (Other Than Silk-Worm Gut)
#>  9:                                            Wood And Articles Of Wood; Wood Charcoal; Cork And Articles Of Cork; Manufactures Of Straw, Of Esparto Or Of Other Plaiting Materials; Basketware And Wickerwork
#> 10:                                                            Pulp Of Wood Or Of Other Fibrous Cellulosic Material; Recovered (Waste And Scrap) Paper Or Paperboard; Paper And Paperboard And Articles Thereof
#> 11:                                                                                                                                                                               Textiles And Textile Articles
#> 12: Footwear, Headgear, Umbrellas, Sun Umbrellas, Walking-Sticks, Seat-Sticks, Whips, Riding-Crops And Parts Thereof; Prepared Feathers And Articles Made Therewith; Artificial Flowers; Articles Of Human Hair
#> 13:                                                                                              Articles Of Stone, Plaster, Cement, Asbestos, Mica Or Similar Materials; Ceramic Products; Glass And Glassware
#> 14:                                              Natural Or Cultured Pearls, Precious Or Semi-Precious Stones, Precious Metals, Metals Clad With Precious Metal And Articles Thereof; Imitation Jewellery; Coin
#> 15:                                                                                                                                                                      Base Metals And Articles Of Base Metal
#> 16: Machinery And Mechanical Appliances; Electrical Equipment; Parts Thereof; Sound Recorders And Reproducers, Television Image And Sound Recorders And Reproducers, And Parts And Accessories Of Such Articles
#> 17:                                                                                                                                              Vehicles, Aircraft, Vessels And Associated Transport Equipment
#> 18:               Optical, Photographic, Cinematographic, Measuring, Checking, Precision, Medical Or Surgical Instruments And Apparatus; Clocks And Watches; Musical Instruments; Parts And Accessories Thereof
#> 19:                                                                                                                                                          Arms And Ammunition; Parts And Accessories Thereof
#> 20:                                                                                                                                                                         Miscellaneous Manufactured Articles
#> 21:                                                                                                                                                               Works Of Art, Collectors' Pieces And Antiques
#> 22:                                                                                                                                                                 Commodities Not Specified According To Kind
#>                                                                                                                                                                                        section_fullname_english