String matching of official country names and ISO-3 codes according to the United Nations nomenclature

Takes a text string and searches within the package data for a country code in the context of valid API country codes.

ots_country_code(countryname = NULL)

Arguments

countryname
A text string such as "Chile", "CHILE" or "CHL".

Value

A single character if there is a exact match (e.g. ots_country_code("Chile")) or a tibble in case of multiple matches (e.g. ots_country_code("Germany"))

Examples

ots_country_code("Chile ")

#>    country_iso country_name_english country_fullname_english continent_id
#> 1:         chl                Chile                    Chile            5
#>    continent eu28_member
#> 1:  Americas           0

ots_country_code("america")

#>    country_iso country_name_english
#> 1:         usa                  USA
#>                                                       country_fullname_english
#> 1: USA, Puerto Rico and US Virgin Islands (excludes Virgin Islands until 1981)
#>    continent_id continent eu28_member
#> 1:            5  Americas           0

ots_country_code("UNITED  STATES")

#> Empty data.table (0 rows and 6 cols): country_iso,country_name_english,country_fullname_english,continent_id,continent,eu28_member

ots_country_code(" united_")

#>    country_iso                 country_name_english
#> 1:         are                 United Arab Emirates
#> 2:         gbr                       United Kingdom
#> 3:         tza              United Rep. of Tanzania
#> 4:         umi United States Minor Outlying Islands
#>                country_fullname_english continent_id continent eu28_member
#> 1:                 United Arab Emirates            1      Asia           0
#> 2:                       United Kingdom            2    Europe           1
#> 3:              United Rep. of Tanzania            3    Africa           0
#> 4: United States Minor Outlying Islands            5  Americas           0