A table with world weigthed mean inflation since 1962

Provides year to year inflations value to be applied as a conversion rate to express dollars of year Y1 as dollars of year Y2. This dataset is provided to be used with ots_inflation_adjustment that converts units forwards and backwards in time.

data("ots_inflation")

Format

A data frame with 56 observations on the following 3 variables.

from

Integer values in the range 1962-2017

to

Integer values in the range 1963-2018

conversion_factor

Numeric value expressed as one plus 1-year inflation

Examples

ots_inflation

#>    from   to conversion_factor
#> 1  1962 1963            1.0308
#> 2  1963 1964            1.0338
#> 3  1964 1965            1.0412
#> 4  1965 1966            1.0532
#> 5  1966 1967            1.0354
#> 6  1967 1968            1.0398
#> 7  1968 1969            1.0364
#> 8  1969 1970            1.0502
#> 9  1970 1971            1.0546
#> 10 1971 1972            1.0588
#> 11 1972 1973            1.0917
#> 12 1973 1974            1.1540
#> 13 1974 1975            1.1247
#> 14 1975 1976            1.0969
#> 15 1976 1977            1.1035
#> 16 1977 1978            1.0862
#> 17 1978 1979            1.1010
#> 18 1979 1980            1.1374
#> 19 1980 1981            1.1284
#> 20 1981 1982            1.1038
#> 21 1982 1983            1.0937
#> 22 1983 1984            1.0904
#> 23 1984 1985            1.0868
#> 24 1985 1986            1.0650
#> 25 1986 1987            1.0698
#> 26 1987 1988            1.0997
#> 27 1988 1989            1.1464
#> 28 1989 1990            1.2329
#> 29 1990 1991            1.1104
#> 30 1991 1992            1.1096
#> 31 1992 1993            1.2079
#> 32 1993 1994            1.2389
#> 33 1994 1995            1.0915
#> 34 1995 1996            1.0651
#> 35 1996 1997            1.0472
#> 36 1997 1998            1.0439
#> 37 1998 1999            1.0342
#> 38 1999 2000            1.0368
#> 39 2000 2001            1.0379
#> 40 2001 2002            1.0295
#> 41 2002 2003            1.0310
#> 42 2003 2004            1.0317
#> 43 2004 2005            1.0363
#> 44 2005 2006            1.0378
#> 45 2006 2007            1.0407
#> 46 2007 2008            1.0695
#> 47 2008 2009            1.0233
#> 48 2009 2010            1.0299
#> 49 2010 2011            1.0450
#> 50 2011 2012            1.0348
#> 51 2012 2013            1.0274
#> 52 2013 2014            1.0228
#> 53 2014 2015            1.0151
#> 54 2015 2016            1.0157
#> 55 2016 2017            1.0241
#> 56 2017 2018            1.0258