String matching of official product names and Harmonized System (HS) codes according to the United Nations nomenclature

This function takes a text string and searches within the package data for all matching product codes in the context of valid API product codes.

ots_product_code(productname = NULL, productgroup = NULL)

Arguments

productname
A text string such as "Animals", "COPPER" or "fruits".

productgroup
A text string such as "meat", "FISH" or "Dairy".

Value

A tibble with all possible matches (no uppercase distinction) showing the product name, product code and corresponding trade classification (e.g. HS92 or SITC)

Examples

ots_product_code(productname = "animals")

#> # A tibble: 55 x 5
#>    commodity_code commodity_fullname_english  group_code group_name type_product
#>    <chr>          <chr>                       <chr>      <chr>      <chr>       
#>  1 010110         Horses, asses, mules and h… 01         Animals; … animals     
#>  2 010190         Horses, asses, mules and h… 01         Animals; … animals     
#>  3 0102           Bovine animals; live        01         Animals; … animals     
#>  4 010210         Bovine animals; live, pure… 01         Animals; … animals     
#>  5 010290         Bovine animals; live, othe… 01         Animals; … animals     
#>  6 010310         Swine; live, pure-bred bre… 01         Animals; … animals     
#>  7 010391         Swine; live, (other than p… 01         Animals; … animals     
#>  8 010392         Swine; live, (other than p… 01         Animals; … animals     
#>  9 0106           Animals, n.e.c. in chapter… 01         Animals; … animals     
#> 10 010631         Animals; live, birds of pr… 01         Animals; … animals     
#> # … with 45 more rows

ots_product_code(productgroup = "fish")

#> # A tibble: 144 x 5
#>    commodity_code commodity_fullname_engl… group_code group_name      type_group
#>    <chr>          <chr>                    <chr>      <chr>           <chr>     
#>  1 0301           Fish; live               03         Fish and crust… fish      
#>  2 030110         Fish; live, ornamental   03         Fish and crust… fish      
#>  3 030191         Fish; live, trout (salm… 03         Fish and crust… fish      
#>  4 030192         Fish; live, eels (angui… 03         Fish and crust… fish      
#>  5 030193         Fish; live, carp         03         Fish and crust… fish      
#>  6 030194         Fish; live, bluefin tun… 03         Fish and crust… fish      
#>  7 030195         Fish; live, southern bl… 03         Fish and crust… fish      
#>  8 030199         Fish; live, n.e.c. in h… 03         Fish and crust… fish      
#>  9 0302           Fish; fresh or chilled … 03         Fish and crust… fish      
#> 10 030211         Fish; trout (salmo trut… 03         Fish and crust… fish      
#> # … with 134 more rows

ots_product_code(productname = "milk", productgroup = "dairy")

#> # A tibble: 17 x 6
#>    commodity_code commodity_fullnam… group_code group_name  type_name type_group
#>    <chr>          <chr>              <chr>      <chr>       <chr>     <chr>     
#>  1 040120         Dairy produce; mi… 04         Dairy prod… milk      dairy     
#>  2 0401           Milk and cream; n… 04         Dairy prod… milk      dairy     
#>  3 040110         Dairy produce; mi… 04         Dairy prod… milk      dairy     
#>  4 040130         Dairy produce; mi… 04         Dairy prod… milk      dairy     
#>  5 0402           Milk and cream; c… 04         Dairy prod… milk      dairy     
#>  6 040210         Dairy produce; mi… 04         Dairy prod… milk      dairy     
#>  7 040221         Dairy produce; mi… 04         Dairy prod… milk      dairy     
#>  8 040229         Dairy produce; mi… 04         Dairy prod… milk      dairy     
#>  9 040291         Dairy produce; mi… 04         Dairy prod… milk      dairy     
#> 10 040299         Dairy produce; mi… 04         Dairy prod… milk      dairy     
#> 11 0403           Buttermilk, curdl… 04         Dairy prod… milk      dairy     
#> 12 040390         Dairy produce; bu… 04         Dairy prod… milk      dairy     
#> 13 0404           Whey and products… 04         Dairy prod… milk      dairy     
#> 14 040490         Dairy produce; na… 04         Dairy prod… milk      dairy     
#> 15 0405           Butter and other … 04         Dairy prod… milk      dairy     
#> 16 040510         Dairy produce; de… 04         Dairy prod… milk      dairy     
#> 17 040590         Dairy produce; fa… 04         Dairy prod… milk      dairy