String matching of official product names and Harmonized System (HS) codes according to the United Nations nomenclature

This function takes a text string and searches within the package data for all matching product codes in the context of valid API product codes.

ots_product_code(productname = NULL, productgroup = NULL)

Arguments

productname
A text string such as "Animals", "COPPER" or "fruits".

productgroup
A text string such as "meat", "FISH" or "Dairy".

Value

A tibble with all possible matches (no uppercase distinction) showing the product name, product code and corresponding trade classification (e.g. HS92 or SITC)

Examples

ots_product_code(productname = "ANIMALS ")

#> # A tibble: 0 x 5
#> # ... with 5 variables: product_code <chr>, product_fullname_english <chr>,
#> #   group_code <chr>, group_name <chr>, type_product <chr>

ots_product_code(productgroup = "  fish")

#> # A tibble: 12 x 5
#>    product_code product_fullname_english group_code group_name        type_group
#>    <chr>        <chr>                    <chr>      <chr>             <chr>     
#>  1 0301         Fish; live               03         Fish and crustac~ fish      
#>  2 0302         Fish; fresh or chilled ~ 03         Fish and crustac~ fish      
#>  3 0303         Fish; frozen (excluding~ 03         Fish and crustac~ fish      
#>  4 0304         Fish fillets and other ~ 03         Fish and crustac~ fish      
#>  5 0305         Fish; dried, salted or ~ 03         Fish and crustac~ fish      
#>  6 0306         Crustaceans, in shell o~ 03         Fish and crustac~ fish      
#>  7 0307         Molluscs, whether in sh~ 03         Fish and crustac~ fish      
#>  8 1601         Sausages and similar pr~ 16         Meat, fish or cr~ fish      
#>  9 1602         Prepared or preserved m~ 16         Meat, fish or cr~ fish      
#> 10 1603         Extracts and juices of ~ 16         Meat, fish or cr~ fish      
#> 11 1604         Prepared or preserved f~ 16         Meat, fish or cr~ fish      
#> 12 1605         Crustaceans, molluscs a~ 16         Meat, fish or cr~ fish      

ots_product_code(productname = "Milk", productgroup = "Dairy")

#> # A tibble: 0 x 6
#> # ... with 6 variables: product_code <chr>, product_fullname_english <chr>,
#> #   group_code <chr>, group_name <chr>, type_name <chr>, type_group <chr>

ots_product_code()

#> # A tibble: 1,320 x 4
#>    product_code product_fullname_english              group_code group_name     
#>    <chr>        <chr>                                 <chr>      <chr>          
#>  1 0101         Horses, asses, mules and hinnies; li~ 01         Animals; live  
#>  2 0102         Bovine animals; live                  01         Animals; live  
#>  3 0103         Swine; live                           01         Animals; live  
#>  4 0104         Sheep and goats; live                 01         Animals; live  
#>  5 0105         Poultry; live, fowls of the species ~ 01         Animals; live  
#>  6 0106         Animals, n.e.c. in chapter 01; live   01         Animals; live  
#>  7 0201         Meat of bovine animals; fresh or chi~ 02         Meat and edibl~
#>  8 0202         Meat of bovine animals; frozen        02         Meat and edibl~
#>  9 0203         Meat of swine; fresh, chilled or fro~ 02         Meat and edibl~
#> 10 0204         Meat of sheep or goats; fresh, chill~ 02         Meat and edibl~
#> # ... with 1,310 more rows