String matching of official product names and Harmonized System (HS) codes according to the United Nations nomenclature

This function takes a text string and searches within the package data for all matching product codes in the context of valid API product codes.

ots_product_code(productname = NULL, productgroup = NULL)

Arguments

productname
A text string such as "Animals", "COPPER" or "fruits".

productgroup
A text string such as "meat", "FISH" or "Dairy".

Value

A tibble with all possible matches (no uppercase distinction) showing the product name, product code and corresponding trade classification (e.g. HS92 or SITC)

Examples

ots_product_code(productname = "ANIMALS ")

#> # A tibble: 0 x 5
#> # … with 5 variables: product_code <chr>, product_fullname_english <chr>,
#> #   group_code <chr>, group_name <chr>, type_product <chr>

ots_product_code(productgroup = "  fish")

#> # A tibble: 144 x 5
#>    product_code product_fullname_english  group_code group_name       type_group
#>    <chr>        <chr>                     <chr>      <chr>            <chr>     
#>  1 0301         Fish; live                03         Fish and crusta… fish      
#>  2 030110       Fish; live, ornamental    03         Fish and crusta… fish      
#>  3 030191       Fish; live, trout (salmo… 03         Fish and crusta… fish      
#>  4 030192       Fish; live, eels (anguil… 03         Fish and crusta… fish      
#>  5 030193       Fish; live, carp          03         Fish and crusta… fish      
#>  6 030194       Fish; live, bluefin tuna… 03         Fish and crusta… fish      
#>  7 030195       Fish; live, southern blu… 03         Fish and crusta… fish      
#>  8 030199       Fish; live, n.e.c. in he… 03         Fish and crusta… fish      
#>  9 0302         Fish; fresh or chilled (… 03         Fish and crusta… fish      
#> 10 030211       Fish; trout (salmo trutt… 03         Fish and crusta… fish      
#> # … with 134 more rows

ots_product_code(productname = "Milk", productgroup = "Dairy")

#> # A tibble: 0 x 6
#> # … with 6 variables: product_code <chr>, product_fullname_english <chr>,
#> #   group_code <chr>, group_name <chr>, type_name <chr>, type_group <chr>

ots_product_code()

#> # A tibble: 6,373 x 4
#>    product_code product_fullname_english                  group_code group_name 
#>    <chr>        <chr>                                     <chr>      <chr>      
#>  1 0101         Horses, asses, mules and hinnies; live    01         Animals; l…
#>  2 010110       Horses, asses, mules and hinnies; live, … 01         Animals; l…
#>  3 010190       Horses, asses, mules and hinnies; live, … 01         Animals; l…
#>  4 0102         Bovine animals; live                      01         Animals; l…
#>  5 010210       Bovine animals; live, pure-bred breeding… 01         Animals; l…
#>  6 010290       Bovine animals; live, other than pure-br… 01         Animals; l…
#>  7 0103         Swine; live                               01         Animals; l…
#>  8 010310       Swine; live, pure-bred breeding animals   01         Animals; l…
#>  9 010391       Swine; live, (other than pure-bred breed… 01         Animals; l…
#> 10 010392       Swine; live, (other than pure-bred breed… 01         Animals; l…
#> # … with 6,363 more rows