String matching of official product section names and product section codes

Takes a text string and searches within the package data for all matching product communities in the context of valid API product communities

ots_product_section(productsection = NULL)

Arguments

productsection
A text string such as "animals", or "FOODSTUFFS".

Value

A tibble with all possible matches (no uppercase distinction) showing the section name and section code

Examples

ots_product_section(productsection = "  Animals")

#> Joining, by = "section_code"

#> # A tibble: 42 x 4
#>    product_code section_code section_fullname_english      type_section
#>    <chr>        <chr>        <chr>                         <chr>       
#>  1 0101         01           Live Animals; Animal Products Animals     
#>  2 0102         01           Live Animals; Animal Products Animals     
#>  3 0103         01           Live Animals; Animal Products Animals     
#>  4 0104         01           Live Animals; Animal Products Animals     
#>  5 0105         01           Live Animals; Animal Products Animals     
#>  6 0106         01           Live Animals; Animal Products Animals     
#>  7 0201         01           Live Animals; Animal Products Animals     
#>  8 0202         01           Live Animals; Animal Products Animals     
#>  9 0203         01           Live Animals; Animal Products Animals     
#> 10 0204         01           Live Animals; Animal Products Animals     
#> # … with 32 more rows

ots_product_section(productsection = "FABRIC ")

#> Joining, by = "section_code"

#> # A tibble: 0 x 4
#> # … with 4 variables: product_code <chr>, section_code <chr>,
#> #   section_fullname_english <chr>, type_section <chr>